Move `MainQueueDispatchDecorator` to separate file
