+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedUIComposer.swift
@@ -26,38 +26,6 @@ public final class FeedUIComposer {
 	}
 }
 
-private final class MainQueueDispatchDecorator<T> {
-	private let decoratee: T
-	
-	init(decoratee: T) {
-		self.decoratee = decoratee
-	}
-	
-	func dispatch(completion: @escaping () -> Void) {
-		guard Thread.isMainThread else {
-			return DispatchQueue.main.async(execute: completion)
-		}
-		
-		completion()
-	}
-}
-
-extension MainQueueDispatchDecorator: FeedLoader where T == FeedLoader {
-	func load(completion: @escaping (FeedLoader.Result) -> Void) {
-		decoratee.load { [weak self] result in
-			self?.dispatch { completion(result) }
-		}
-	}
-}
-
-extension MainQueueDispatchDecorator: FeedImageDataLoader where T == FeedImageDataLoader {
-	func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-		return decoratee.loadImageData(from: url) { [weak self] result in
-			self?.dispatch { completion(result) }
-		}
-	}
-}
-
 private extension FeedViewController {
 	static func makeWith(delegate: FeedViewControllerDelegate, title: String) -> FeedViewController {
 		let bundle = Bundle(for: FeedViewController.self)
